+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@3.2.2...@cognite/sdk-core@3.2.3) (2021-08-20)
+
+**Note:** Version bump only for package @cognite/sdk-core
+
+
+
+
+
 ## [3.2.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@3.2.1...@cognite/sdk-core@3.2.2) (2021-08-11)